Nutrilens_Model/backend/requirements.txt
fastapi
uvicorn
ultralytics
python-multipart
opencv-python-headless
numpy
pillow
pydantic
pydantic-settings
python-dotenv
huggingface-hub
pyyaml
openai
